+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Drop support for PHP 7.2 and 7.3
 * Require `guzzlehttp/command` ^2.0, `guzzlehttp/guzzle` ^8.0, `guzzlehttp/psr7` ^3.0, and `guzzlehttp/uri-template` ^2.0
 * Default operations without an `httpMethod` to `GET` and reject invalid `httpMethod` values
+* Require header location values to be strings or arrays of strings
 
 ## 1.6.0 - UPCOMING